src/requirements.txt
python-dotenv

# Django
django
whitenoise

# Databases
dj-database-url

# Telegram
python-telegram-bot[passport,json]
pywaves
pycryptodome
pynacl